pyproject.toml
[tool.poetry]
name = "chatavd"
version = "0.1.0"
description = ""
authors = ["Fred Hsu <fredlhsu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-fasthtml = "^0.6.4"
langchain = "^0.3.0"
langchain-core = "^0.3.2"
langchain-openai = "^0.2.0"
langchain-chroma = "^0.1.4"
langchain-qdrant = "^0.1.4"
libsql-experimental = "^0.0.40"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
